+++ lessons/01-vue-js-create-rxjs-streams-in-the-vue-js-subscriptions-function/src/App.vue
@@ -1,24 +1,24 @@
 <template>
   <section class="section">
-    <h1 class="title">{{timesTwo$}}</h1>
-    <h1 class="title">{{timesThree$}}</h1>
+    <h1 class="title">{{ timesTwo$ }}</h1>
+    <h1 class="title">{{ timesThree$ }}</h1>
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { interval } from "rxjs";
+import { map } from "rxjs/operators";
 
 export default {
   subscriptions() {
-    const interval$ = Observable.interval(1000)
+    const interval$ = interval(1000);
 
-    const timesTwo$ = interval$.map(i => i * 2)
-    const timesThree$ = interval$.map(i => i * 3)
+    const timesTwo$ = interval$.pipe(map(i => i * 2));
+    const timesThree$ = interval$.pipe(map(i => i * 3));
     return {
       timesTwo$,
       timesThree$
-    }
+    };
   }
-}
+};
 </script>
-