+++ lessons/01-vue-js-create-rxjs-streams-in-the-vue-js-subscriptions-function/src/main.js
@@ -3,12 +3,11 @@ import 'buefy/dist/buefy.css'
 import Vue from "vue"
 import App from "./App.vue"
 
-import Rx from "rxjs"
 import VueRx from "vue-rx"
 
 import Buefy from "buefy"
 
-Vue.use(VueRx, Rx)
+Vue.use(VueRx)
 Vue.use(Buefy)
 
 Vue.config.productionTip = false