+++ lessons/02-vue-js-access-events-from-vue-js-templates-as-rxjs-streams-with-domstreams/src/App.vue
@@ -1,24 +1,22 @@
 <template>
   <section class="section">
     <button class="button" v-stream:click="click$">Click</button>
-    <h1 class="title">{{random$}}</h1>
+    <h1 class="title">{{ random$ }}</h1>
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { Observable } from "rxjs";
+import { map } from "rxjs/operators";
 
 export default {
   domStreams: ["click$"],
   subscriptions() {
-    const random$ = this.click$.map(() =>
-      Math.random()
-    )
+    const random$ = this.click$.pipe(map(() => Math.random()));
 
     return {
       random$
-    }
+    };
   }
-}
+};
 </script>
-