+++ lessons/02-vue-js-access-events-from-vue-js-templates-as-rxjs-streams-with-domstreams/src/main.js
@@ -1,14 +1,13 @@
-import "buefy/lib/buefy.css"
+import 'buefy/dist/buefy.css'
 
 import Vue from "vue"
 import App from "./App.vue"
 
-import Rx from "rxjs"
 import VueRx from "vue-rx"
 
 import Buefy from "buefy"
 
-Vue.use(VueRx, Rx)
+Vue.use(VueRx)
 Vue.use(Buefy)
 
 Vue.config.productionTip = false