+++ lessons/03-vue-js-stream-an-api-using-rxjs-into-a-vue-js-template/src/App.vue
@@ -1,31 +1,27 @@
 <template>
   <section class="section">
     <button class="button" v-stream:click="click$">Click</button>
-    <h1 class="title">{{people$}}</h1>
+    <h1 class="title">{{ people$ }}</h1>
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { from } from "rxjs";
+import { map, pluck } from "rxjs/operators";
 
 export default {
   domStreams: ["click$"],
   subscriptions() {
-    const people$ = Observable.from(
-      this.$http.get(
-        "https://starwars.egghead.training/people/1"
-      )
-    ).pluck("data", "name")
+    const people$ = from(
+      this.$http.get("https://starwars.egghead.training/people/1")
+    ).pipe(pluck("data", "name"));
 
-    const random$ = this.click$.map(() =>
-      Math.random()
-    )
+    const random$ = this.click$.pipe(map(() => Math.random()));
 
     return {
       random$,
       people$
-    }
+    };
   }
-}
+};
 </script>
-