+++ lessons/04-vue-js-map-click-event-streams-to-load-data-with-vue-js-and-rxjs/src/App.vue
@@ -1,30 +1,26 @@
 <template>
   <section class="section">
     <button class="button" v-stream:click="click$">Click</button>
-    <h1 class="title">{{luke$}}</h1>
+    <h1 class="title">{{ luke$ }}</h1>
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { from } from "rxjs";
+import { pluck, switchMap } from "rxjs/operators";
 
 export default {
   domStreams: ["click$"],
   subscriptions() {
-    const people$ = Observable.from(
-      this.$http.get(
-        "https://starwars.egghead.training/people/1"
-      )
-    ).pluck("data", "name")
+    const people$ = from(
+      this.$http.get("https://starwars.egghead.training/people/1")
+    ).pipe(pluck("data", "name"));
 
-    const luke$ = this.click$.switchMap(
-      () => people$
-    )
+    const luke$ = this.click$.pipe(switchMap(() => people$));
 
     return {
       luke$
-    }
+    };
   }
-}
+};
 </script>
-