+++ lessons/05-vue-js-split-a-request-into-data-stream-and-an-image-stream-with-rxjs-and-vue-js/src/App.vue
@@ -1,40 +1,34 @@
 <template>
   <section class="section">
     <button class="button" v-stream:click="click$">Click</button>
-    <h1 class="title">{{name$}}</h1>
-    <img :src="image$" alt="">
+    <h1 class="title">{{ name$ }}</h1>
+    <img :src="image$" alt>
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { from } from "rxjs";
+import { pluck, map, switchMap } from "rxjs/operators";
 
 export default {
   domStreams: ["click$"],
   subscriptions() {
-    const person$ = Observable.from(
-      this.$http.get(
-        "https://starwars.egghead.training/people/1"
-      )
-    ).pluck("data")
+    const person$ = from(
+      this.$http.get("https://starwars.egghead.training/people/1")
+    ).pipe(pluck("data"));
 
-    const luke$ = this.click$.switchMap(
-      () => person$
-    )
+    const luke$ = this.click$.pipe(switchMap(() => person$));
 
-    const name$ = luke$.pluck("name")
-    const image$ = luke$
-      .pluck("image")
-      .map(
-        image =>
-          `https://starwars.egghead.training/${image}`
-      )
+    const name$ = luke$.pipe(pluck("name"));
+    const image$ = luke$.pipe(
+      pluck("image"),
+      map(image => `https://starwars.egghead.training/${image}`)
+    );
 
     return {
       name$,
       image$
-    }
+    };
   }
-}
+};
 </script>
-