+++ lessons/06-vue-js-handle-image-loading-errors-in-vue-js-with-rxjs-and-domstreams/src/App.vue
@@ -1,48 +1,38 @@
 <template>
   <section class="section">
     <button class="button" v-stream:click="click$">Click</button>
-    <h1 class="title">{{name$}}</h1>
-    <img v-stream:error="imageError$" :src="image$" alt="">
+    <h1 class="title">{{ name$ }}</h1>
+    <img v-stream:error="imageError$" :src="image$" alt />
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { from, merge } from "rxjs";
+import { pluck, switchMap, map, mapTo } from "rxjs/operators";
 
 export default {
   domStreams: ["click$", "imageError$"],
   subscriptions() {
-    const person$ = Observable.from(
-      this.$http.get(
-        "https://starwars.egghead.training/people/1"
-      )
-    ).pluck("data")
+    const person$ = from(
+      this.$http.get("https://starwars.egghead.training/people/1")
+    ).pipe(pluck("data"));
 
-    const luke$ = this.click$.switchMap(
-      () => person$
-    )
+    const luke$ = this.click$.pipe(switchMap(() => person$));
 
-    const name$ = luke$.pluck("name")
+    const name$ = luke$.pipe(pluck("name"));
     const loadImage$ = luke$
-      .pluck("image")
-      .map(
-        image =>
-          `https://starwars.egghead.training/${image}`
-      )
+      .pipe(pluck("image"))
+      .pipe(map(image => `https://starwars.egghead.training/${image}`));
 
-    const failImage$ = this.imageError$.mapTo(
-      `http://via.placeholder.com/400x400`
-    )
+    const failImage$ = this.imageError$.pipe(
+      mapTo(`http://via.placeholder.com/400x400`)
+    );
 
-    const image$ = Observable.merge(
-      loadImage$,
-      failImage$
-    )
+    const image$ = merge(loadImage$, failImage$);
     return {
       name$,
       image$
-    }
+    };
   }
-}
+};
 </script>
-