+++ lessons/07-vue-js-switch-to-a-function-which-creates-observables-with-vue-js-and-rxjs/src/App.vue
@@ -1,49 +1,40 @@
 <template>
   <section class="section">
     <button class="button" v-stream:click="click$">Click</button>
-    <h1 class="title">{{name$}}</h1>
-    <img v-stream:error="imageError$" :src="image$" alt="">
+    <h1 class="title">{{ name$ }}</h1>
+    <img v-stream:error="imageError$" :src="image$" alt>
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { from, merge } from "rxjs";
+import { pluck, mapTo, switchMap, map } from "rxjs/operators";
 
 export default {
   domStreams: ["click$", "imageError$"],
   subscriptions() {
-    const createLoader = url =>
-      Observable.from(this.$http.get(url)).pluck(
-        "data"
-      )
+    const createLoader = url => from(this.$http.get(url)).pipe(pluck("data"));
 
-    const luke$ = this.click$
-      .mapTo(
-        "https://starwars.egghead.training/people/1"
-      )
-      .switchMap(createLoader)
+    const luke$ = this.click$.pipe(
+      mapTo("https://starwars.egghead.training/people/1"),
+      switchMap(createLoader)
+    );
 
-    const name$ = luke$.pluck("name")
-    const loadImage$ = luke$
-      .pluck("image")
-      .map(
-        image =>
-          `https://starwars.egghead.training/${image}`
-      )
+    const name$ = luke$.pipe(pluck("name"));
+    const loadImage$ = luke$.pipe(
+      pluck("image"),
+      map(image => `https://starwars.egghead.training/${image}`)
+    );
 
-    const failImage$ = this.imageError$.mapTo(
-      `http://via.placeholder.com/400x400`
-    )
+    const failImage$ = this.imageError$.pipe(
+      mapTo(`http://via.placeholder.com/400x400`)
+    );
 
-    const image$ = Observable.merge(
-      loadImage$,
-      failImage$
-    )
+    const image$ = merge(loadImage$, failImage$);
     return {
       name$,
       image$
-    }
+    };
   }
-}
+};
 </script>
-