+++ lessons/07-vue-js-switch-to-a-function-which-creates-observables-with-vue-js-and-rxjs/src/main.js
@@ -3,7 +3,6 @@ import 'buefy/dist/buefy.css'
 import Vue from "vue"
 import App from "./App.vue"
 
-import Rx from "rxjs"
 import VueRx from "vue-rx"
 
 import Buefy from "buefy"
@@ -11,7 +10,7 @@ import Buefy from "buefy"
 import axios from "axios"
 import VueAxios from "vue-axios"
 
-Vue.use(VueRx, Rx)
+Vue.use(VueRx)
 Vue.use(Buefy)
 Vue.use(VueAxios, axios)