+++ lessons/08-vue-js-catch-errors-when-data-fails-to-load-with-vue-js-and-rxjs/src/main.js
@@ -1,22 +1,21 @@
-import 'buefy/dist/buefy.css'
+import "buefy/dist/buefy.css";
 
-import Vue from "vue"
-import App from "./App.vue"
+import Vue from "vue";
+import App from "./App.vue";
 
-import Rx from "rxjs"
-import VueRx from "vue-rx"
+import VueRx from "vue-rx";
 
-import Buefy from "buefy"
+import Buefy from "buefy";
 
-import axios from "axios"
-import VueAxios from "vue-axios"
+import axios from "axios";
+import VueAxios from "vue-axios";
 
-Vue.use(VueRx, Rx)
-Vue.use(Buefy)
-Vue.use(VueAxios, axios)
+Vue.use(VueRx);
+Vue.use(Buefy);
+Vue.use(VueAxios, axios);
 
-Vue.config.productionTip = false
+Vue.config.productionTip = false;
 
 new Vue({
   render: h => h(App)
-}).$mount("#app")
+}).$mount("#app");