+++ lessons/10-vue-js-disable-buttons-while-data-is-loading-with-rxjs-and-vue-js/src/main.js
@@ -3,7 +3,7 @@ import 'buefy/dist/buefy.css'
 import Vue from "vue"
 import App from "./App.vue"
 
-import Rx from "rxjs"
+
 import VueRx from "vue-rx"
 
 import Buefy from "buefy"
@@ -11,7 +11,7 @@ import Buefy from "buefy"
 import axios from "axios"
 import VueAxios from "vue-axios"
 
-Vue.use(VueRx, Rx)
+Vue.use(VueRx)
 Vue.use(Buefy)
 Vue.use(VueAxios, axios)