+++ lessons/12-vue-js-pass-template-data-through-domstreams-in-vue-js-and-rxjs/src/App.vue
@@ -1,70 +1,79 @@
 <template>
   <section class="section">
-    <button class="button" :disabled="disabled$" v-stream:click="{subject:click$, data:1}">{{buttonText$}}</button>
-    <button class="button" :disabled="disabled$" v-stream:click="{subject:click$, data:4}">{{buttonText$}}</button>
-    <button class="button" :disabled="disabled$" v-stream:click="{subject:click$, data:5}">{{buttonText$}}</button>
-    <h1 class="title">{{name$}}</h1>
-    <img v-stream:error="imageError$" :src="image$" alt="">
+    <button
+      class="button"
+      :disabled="disabled$"
+      v-stream:click="{ subject: click$, data: 1 }"
+    >{{ buttonText$ }}</button>
+    <button
+      class="button"
+      :disabled="disabled$"
+      v-stream:click="{ subject: click$, data: 4 }"
+    >{{ buttonText$ }}</button>
+    <button
+      class="button"
+      :disabled="disabled$"
+      v-stream:click="{ subject: click$, data: 5 }"
+    >{{ buttonText$ }}</button>
+    <h1 class="title">{{ name$ }}</h1>
+    <img v-stream:error="imageError$" :src="image$" alt>
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { from, merge } from "rxjs";
+import {
+  pluck,
+  mapTo,
+  switchMap,
+  map,
+  catchError,
+  share,
+  startWith,
+  exhaustMap
+} from "rxjs/operators";
 
 export default {
   domStreams: ["click$", "imageError$"],
   subscriptions() {
-    const createLoader = url =>
-      Observable.from(this.$http.get(url)).pluck(
-        "data"
-      )
+    const createLoader = url => from(this.$http.get(url)).pipe(pluck("data"));
 
-    const luke$ = this.click$
-      .pluck("data")
-      .map(
-        id =>
-          `https://starwars.egghead.training/people/${id}`
-      )
-      .exhaustMap(createLoader)
-      .catch(err =>
-        createLoader(
-          "https://starwars.egghead.training/people/2"
-        )
-      )
-      .share()
+    const luke$ = this.click$.pipe(
+      pluck("data"),
+      map(id => `https://starwars.egghead.training/people/${id}`),
+      exhaustMap(createLoader),
+      catchError(err =>
+        createLoader("https://starwars.egghead.training/people/2")
+      ),
+      share()
+    );
 
-    const disabled$ = Observable.merge(
-      this.click$.mapTo(true),
-      luke$.mapTo(false)
-    ).startWith(false)
+    const disabled$ = merge(
+      this.click$.pipe(mapTo(true)),
+      luke$.pipe(mapTo(false))
+    ).pipe(startWith(false));
 
-    const buttonText$ = disabled$.map(
-      bool => (bool ? "Loading..." : "Load")
-    )
+    const buttonText$ = disabled$.pipe(
+      map(bool => (bool ? "Loading..." : "Load"))
+    );
 
-    const name$ = luke$.pluck("name")
-    const loadImage$ = luke$
-      .pluck("image")
-      .map(
-        image =>
-          `https://starwars.egghead.training/${image}`
-      )
+    const name$ = luke$.pipe(pluck("name"));
+    const loadImage$ = luke$.pipe(
+      pluck("image"),
+      map(image => `https://starwars.egghead.training/${image}`)
+    );
 
-    const failImage$ = this.imageError$.mapTo(
-      `http://via.placeholder.com/400x400`
-    )
+    const failImage$ = this.imageError$.pipe(
+      mapTo(`http://via.placeholder.com/400x400`)
+    );
 
-    const image$ = Observable.merge(
-      loadImage$,
-      failImage$
-    )
+    const image$ = merge(loadImage$, failImage$);
     return {
       name$,
       image$,
       disabled$,
       buttonText$
-    }
+    };
   }
-}
+};
 </script>
-