+++ lessons/14-vue-js-sync-vue-js-components-with-streaming-content-using-watchasobservable-and-rxjs/src/App.vue
@@ -1,16 +1,20 @@
 <template>
   <section class="section">
     <b-tabs v-model="activeTab">
-      <b-tab-item v-for="person of people" :key="person.id" :label="person.name"></b-tab-item>
+      <b-tab-item
+        v-for="person of people"
+        :key="person.id"
+        :label="person.name"
+      ></b-tab-item>
     </b-tabs>
 
-    <h1 class="title">{{name$}}</h1>
-    <img v-stream:error="imageError$" :src="image$" alt="">
+    <h1 class="title">{{ name$ }}</h1>
+    <img v-stream:error="imageError$" :src="image$" alt="" />
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { Observable } from "rxjs";
 
 export default {
   data() {
@@ -22,67 +26,48 @@ export default {
         { name: "Leia", id: 5 },
         { name: "Yoda", id: 20 }
       ]
-    }
+    };
   },
   domStreams: ["click$", "imageError$"],
   subscriptions() {
-    const activeTab$ = this.$watchAsObservable(
-      "activeTab",
-      { immediate: true }
-    ).pluck("newValue")
+    const activeTab$ = this.$watchAsObservable("activeTab", {
+      immediate: true
+    }).pluck("newValue");
 
     const createLoader = url =>
-      Observable.from(this.$http.get(url)).pluck(
-        "data"
-      )
+      Observable.from(this.$http.get(url)).pluck("data");
 
     const luke$ = activeTab$
       .map(tabId => this.people[tabId].id)
-      .map(
-        id =>
-          `https://starwars.egghead.training/people/${id}`
-      )
+      .map(id => `https://starwars.egghead.training/people/${id}`)
       .exhaustMap(createLoader)
-      .catch(err =>
-        createLoader(
-          "https://starwars.egghead.training/people/2"
-        )
-      )
-      .share()
+      .catch(err => createLoader("https://starwars.egghead.training/people/2"))
+      .share();
 
     const disabled$ = Observable.merge(
       this.click$.mapTo(true),
       luke$.mapTo(false)
-    ).startWith(false)
+    ).startWith(false);
 
-    const buttonText$ = disabled$.map(
-      bool => (bool ? "Loading..." : "Load")
-    )
+    const buttonText$ = disabled$.map(bool => (bool ? "Loading..." : "Load"));
 
-    const name$ = luke$.pluck("name")
+    const name$ = luke$.pluck("name");
     const loadImage$ = luke$
       .pluck("image")
-      .map(
-        image =>
-          `https://starwars.egghead.training/${image}`
-      )
+      .map(image => `https://starwars.egghead.training/${image}`);
 
     const failImage$ = this.imageError$.mapTo(
       `http://via.placeholder.com/400x400`
-    )
+    );
 
-    const image$ = Observable.merge(
-      loadImage$,
-      failImage$
-    )
+    const image$ = Observable.merge(loadImage$, failImage$);
     return {
       name$,
       image$,
       disabled$,
       buttonText$,
       activeTab$
-    }
+    };
   }
-}
+};
 </script>
-