+++ lessons/15-vue-js-map-vue-js-components-to-remote-data-streams-with-rxjs/src/App.vue
@@ -1,81 +1,64 @@
 <template>
   <section class="section">
     <b-tabs v-model="activeTab">
-      <b-tab-item v-for="person of people$" :key="person.id" :label="person.name"></b-tab-item>
+      <b-tab-item
+        v-for="person of people$"
+        :key="person.id"
+        :label="person.name"
+      ></b-tab-item>
     </b-tabs>
 
-    <h1 class="title">{{name$}}</h1>
-    <img v-stream:error="imageError$" :src="image$" alt="">
+    <h1 class="title">{{ name$ }}</h1>
+    <img v-stream:error="imageError$" :src="image$" alt="" />
   </section>
 </template>
 
 <script>
-import { Observable } from "rxjs"
+import { Observable } from "rxjs";
 
 export default {
   data() {
     return {
       activeTab: 0
-    }
+    };
   },
   domStreams: ["click$", "imageError$"],
   subscriptions() {
     const createLoader = url =>
-      Observable.from(this.$http.get(url)).pluck(
-        "data"
-      )
+      Observable.from(this.$http.get(url)).pluck("data");
 
     const people$ = createLoader(
       `https://starwars.egghead.training/people`
-    ).map(people => people.slice(0, 7))
+    ).map(people => people.slice(0, 7));
 
-    const activeTab$ = this.$watchAsObservable(
-      "activeTab",
-      { immediate: true }
-    ).pluck("newValue")
+    const activeTab$ = this.$watchAsObservable("activeTab", {
+      immediate: true
+    }).pluck("newValue");
 
     const luke$ = activeTab$
-      .combineLatest(
-        people$,
-        (tabId, people) => people[tabId].id
-      )
-      .map(
-        id =>
-          `https://starwars.egghead.training/people/${id}`
-      )
+      .combineLatest(people$, (tabId, people) => people[tabId].id)
+      .map(id => `https://starwars.egghead.training/people/${id}`)
       .exhaustMap(createLoader)
-      .catch(err =>
-        createLoader(
-          "https://starwars.egghead.training/people/2"
-        )
-      )
-      .share()
+      .catch(err => createLoader("https://starwars.egghead.training/people/2"))
+      .share();
 
     const disabled$ = Observable.merge(
       this.click$.mapTo(true),
       luke$.mapTo(false)
-    ).startWith(false)
+    ).startWith(false);
 
-    const buttonText$ = disabled$.map(
-      bool => (bool ? "Loading..." : "Load")
-    )
+    const buttonText$ = disabled$.map(bool => (bool ? "Loading..." : "Load"));
 
-    const name$ = luke$.pluck("name")
+    const name$ = luke$.pluck("name");
     const loadImage$ = luke$
       .pluck("image")
-      .map(
-        image =>
-          `https://starwars.egghead.training/${image}`
-      )
+      .map(image => `https://starwars.egghead.training/${image}`);
 
     const failImage$ = this.imageError$.mapTo(
       `http://via.placeholder.com/400x400`
-    )
+    );
 
-    const image$ = Observable.merge(
-      loadImage$,
-      failImage$
-    )
+    const image$ = Observable.merge(loadImage$, failImage$);
     return {
       name$,
       image$,
@@ -83,8 +66,7 @@ export default {
       buttonText$,
       activeTab$,
       people$
-    }
+    };
   }
-}
+};
 </script>
-